% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_haversine.R
\name{distance_haversine}
\alias{distance_haversine}
\title{compute distance in km between two points: http://en.wikipedia.org/wiki/Haversine_formula}
\usage{
distance_haversine(lat1, lon1, lat2, lon2, r = 6378.137)
}
\arguments{
\item{lat1}{latitude of the first geographical location in decimal degrees (WGS84)}

\item{lon1}{longitude of the first geographical location in decimal degrees (WGS84)}

\item{lat2}{latitude of the second geographical location in decimal degrees (WGS84)}

\item{lon2}{longitude of the second geographical location in decimal degrees (WGS84)}

\item{r}{earth diameter at the equator in km}
}
\value{
a numeric value of the distance between the two points in km
}
\description{
compute distance in km between two points: http://en.wikipedia.org/wiki/Haversine_formula
}
\examples{
distance_haversine(43.5279, 1.5009, 42.8504, -0.4333)
}
