% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_penman_monteith.R
\name{et_penman_monteith}
\alias{et_penman_monteith}
\title{Compute reference evapotranspiration according to Penman-Monteith formula and [@Wallach2006] assumption}
\usage{
et_penman_monteith(lat, lon, elevation, day, tmin, tmax, tdew, rad, wind)
}
\arguments{
\item{lat}{Latitude of the geographical location in decimal degrees (WGS84)}

\item{lon}{Longitude of the geographical location in decimal degrees (WGS84)}

\item{elevation}{Elevation of the geographical location in meters (m)}

\item{day}{Day of year used to compute solar irradiation factor (calendar day)}

\item{tmin}{Minimum Air Temperature At 2 m Above The Surface Of The Earth (degrees C)}

\item{tmax}{Maximum Air Temperature At 2 m Above The Surface Of The Earth (degrees C)}

\item{tdew}{Dew/Frost Point Temperature At 2 m (degrees C)}

\item{rad}{Daily Insolation Incident On A Horizontal Surface (MJ/m^2/day)}

\item{wind}{Wind Speed At 10 m Above The Surface Of The Earth (m/s)}
}
\value{
a numeric value for daily reference evapotranspiration (mm/day)
}
\description{
Compute reference evapotranspiration according to Penman-Monteith formula and [@Wallach2006] assumption
}
