% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate.R
\name{get_climate_nasapower}
\alias{get_climate_nasapower}
\title{get climate data from the NASA POWER API using the R nasapower package}
\usage{
get_climate_nasapower(lat, lon, year)
}
\arguments{
\item{lat}{latitude of the geographical location in decimal degrees (WGS84)}

\item{lon}{longitude of the geographical location in decimal degrees (WGS84)}

\item{year}{year(s) of the climate data to be retrieved}
}
\description{
get climate data from the NASA POWER API using the R nasapower package
}
