% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate.R
\name{get_climate_safran}
\alias{get_climate_safran}
\title{get climate data from the SAFRAN model, using a local repository of grid cell data archives}
\usage{
get_climate_safran(lat, lon, year, path)
}
\arguments{
\item{lat}{latitude of the geographical location in decimal degrees (WGS84)}

\item{lon}{longitude of the geographical location in decimal degrees (WGS84)}

\item{year}{year(s) of the climate data to be retrieved}

\item{path}{path to a directory containning grid cell data archives as made available by Meteo-France for research usage}
}
\value{
a dataframe of daily climate data for 9 variables
}
\description{
get climate data from the SAFRAN model, using a local repository of grid cell data archives
}
