% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil.R
\name{get_soil_esdb}
\alias{get_soil_esdb}
\title{get soil data from the European Soil DataBase (ESDB), using a local repository for GIS files.}
\usage{
get_soil_esdb(lat, lon, data = data_soil)
}
\arguments{
\item{lat}{latitude of the geographical location in decimal degrees (WGS84)}

\item{lon}{longitude of the geographical location in decimal degrees (WGS84)}

\item{data}{dataframe created from ESDB GIS files as made available by the JRC lab for research usage.}
}
\value{
a dataframe of soil properties
}
\description{
get soil data from the European Soil DataBase (ESDB), using a local repository for GIS files.
}
