% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station.R
\name{get_station}
\alias{get_station}
\title{get the closest climate station id based on physical station network or grid cell index}
\usage{
get_station(
  latitude,
  longitude,
  data = agroenv::station_index,
  network,
  n = 1,
  ...
)
}
\arguments{
\item{latitude}{latitude of the geographical location in decimal degrees (WGS84)}

\item{longitude}{logitude of the geographical location in decimal degrees (WGS84)}

\item{data}{list of dataframes containing stations or grid cell coordinates}

\item{network}{string of the stations network or gridded data to use}

\item{n}{number of results to return}

\item{...}{additional arguments for use with pmap}
}
\value{
a dataframe of station(s) or grid cell closest to the geographical location
}
\description{
get the closest climate station id based on physical station network or grid cell index
}
\examples{
\donttest{
get_station(43.5279, 1.5009, network="climbox", n=5)
}
}
