% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_index.R
\docType{data}
\name{station_index}
\alias{station_index}
\title{Index for climate station networks or grid cells}
\format{A list of 4 dataframes:
\describe{
  \item{station_id}{id of the station or grid cell}
  \item{lat}{latitude in decimal degrees WGS84}
  \item{lon}{longitude in decimal degrees WGS84}
}}
\source{
\url{http://www/}
}
\usage{
station_index
}
\description{
A list containing index and geographical coordinates of two climate station networks and two grid cells.
}
\keyword{datasets}
